+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] (2.2.0)
+### Removed
+- **Paste and Send** action (`Ctrl+Shift+V`).
+
 ## [2.1.0] - 2022-11-19
 ### Fixed
 - [#50: Crash when opening a hyperlink](https://github.com/ForNeVeR/CyclopsChat/issues/50).